Scripted Jenkinsfile

+++ Jenkinsfile
@@ -31,19 +31,13 @@ node {
     input(message:"Lanjutkan ke tahap Deploy?")
   }
   stage('Deploy') {
-    // try {
-    //   docker.image('cdrx/pyinstaller-linux:python2').inside {
-    //     sh 'pyinstaller --onefile sources/add2vals.py'
-    //   }
-    // } finally {
-    //   if (currentBuild == 'SUCCESS') {
-    //     echo 'This will always run'
-    //     // archiveArtifacts 'dist/add2vals'
-    //   }
-    // }
-    sh 'docker run --rm -v /var/jenkins_home/workspace/submission-cicd-pipeline-ricky_ritonga/sources:/src cdrx/pyinstaller-linux:python2 \'pyinstaller -F add2vals.py\''
-	  archiveArtifacts artifacts: 'sources/add2vals.py', followSymlinks: false
-	  sh 'docker run --rm -v /var/jenkins_home/workspace/submission-cicd-pipeline-ricky_ritonga/sources:/src cdrx/pyinstaller-linux:python2 \'rm -rf build dist\''
-    echo 'This will always run'
+    try {
+      sh 'docker run --rm -v /var/jenkins_home/workspace/submission-cicd-pipeline-ricky_ritonga/sources:/src cdrx/pyinstaller-linux:python2 \'pyinstaller -F add2vals.py\''
+    } finally {
+      if (currentBuild == 'SUCCESS') {
+        archiveArtifacts artifacts: 'sources/add2vals.py', followSymlinks: false
+	      sh 'docker run --rm -v /var/jenkins_home/workspace/submission-cicd-pipeline-ricky_ritonga/sources:/src cdrx/pyinstaller-linux:python2 \'rm -rf build dist\''
+      }
+    }  
   }
 }
